requirements.txt
Django==2.1.3
Markdown==3.0.1
markdown-djangostaticimage==0.1
pytz==2018.7
